# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InflationWeaver",
    "description": "Explicit-input inflation and real-return analytics. Demo data is synthetic.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/catalog": {
      "get": {
        "summary": "Catalog",
        "operationId": "catalog_v1_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/series": {
      "get": {
        "summary": "List Series",
        "operationId": "list_series_v1_series_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/series/{series_id}": {
      "get": {
        "summary": "Get Series",
        "operationId": "get_series_v1_series__series_id__get",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze": {
      "post": {
        "summary": "Analyze Endpoint",
        "operationId": "analyze_endpoint_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/compare": {
      "post": {
        "summary": "Compare Endpoint",
        "operationId": "compare_endpoint_v1_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/portfolio": {
      "post": {
        "summary": "Portfolio Endpoint",
        "operationId": "portfolio_endpoint_v1_portfolio_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cpi/chain": {
      "post": {
        "summary": "Chain Endpoint",
        "operationId": "chain_endpoint_v1_cpi_chain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cpi/splice": {
      "post": {
        "summary": "Splice Endpoint",
        "operationId": "splice_endpoint_v1_cpi_splice_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpliceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/backtest": {
      "post": {
        "summary": "Backtest Endpoint",
        "operationId": "backtest_endpoint_v1_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "base_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Date"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "alignment": {
            "type": "string",
            "enum": [
              "observation",
              "released"
            ],
            "title": "Alignment",
            "default": "observation"
          },
          "max_staleness_days": {
            "type": "integer",
            "maximum": 3660.0,
            "minimum": 0.0,
            "title": "Max Staleness Days",
            "default": 62
          },
          "asset": {
            "$ref": "#/components/schemas/SeriesPayload"
          },
          "inflation": {
            "$ref": "#/components/schemas/SeriesPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "asset",
          "inflation"
        ],
        "title": "AnalyzeRequest"
      },
      "BacktestRequest": {
        "properties": {
          "base_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Date"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "alignment": {
            "type": "string",
            "enum": [
              "observation",
              "released"
            ],
            "title": "Alignment",
            "default": "observation"
          },
          "max_staleness_days": {
            "type": "integer",
            "maximum": 3660.0,
            "minimum": 0.0,
            "title": "Max Staleness Days",
            "default": 62
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/BacktestRow"
            },
            "type": "array",
            "maxItems": 50000,
            "minItems": 2,
            "title": "Data"
          },
          "inflation": {
            "$ref": "#/components/schemas/SeriesPayload"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "TRY"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "inflation"
        ],
        "title": "BacktestRequest"
      },
      "BacktestRow": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "available_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Date"
          },
          "cashflow": {
            "type": "number",
            "title": "Cashflow",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "value"
        ],
        "title": "BacktestRow"
      },
      "ChainRequest": {
        "properties": {
          "rates": {
            "$ref": "#/components/schemas/SeriesPayload"
          },
          "anchor_date": {
            "type": "string",
            "format": "date",
            "title": "Anchor Date"
          },
          "anchor_value": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Anchor Value",
            "default": 100
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rates",
          "anchor_date"
        ],
        "title": "ChainRequest"
      },
      "CompareRequest": {
        "properties": {
          "base_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Date"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "alignment": {
            "type": "string",
            "enum": [
              "observation",
              "released"
            ],
            "title": "Alignment",
            "default": "observation"
          },
          "max_staleness_days": {
            "type": "integer",
            "maximum": 3660.0,
            "minimum": 0.0,
            "title": "Max Staleness Days",
            "default": 62
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/SeriesPayload"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Assets"
          },
          "inflation": {
            "$ref": "#/components/schemas/SeriesPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "assets",
          "inflation"
        ],
        "title": "CompareRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Observation": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "available_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "value"
        ],
        "title": "Observation"
      },
      "PortfolioRequest": {
        "properties": {
          "base_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Date"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "alignment": {
            "type": "string",
            "enum": [
              "observation",
              "released"
            ],
            "title": "Alignment",
            "default": "observation"
          },
          "max_staleness_days": {
            "type": "integer",
            "maximum": 3660.0,
            "minimum": 0.0,
            "title": "Max Staleness Days",
            "default": 62
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/SeriesPayload"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Assets"
          },
          "weights": {
            "items": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Weights"
          },
          "initial_value": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Initial Value",
            "default": 100
          },
          "inflation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeriesPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "assets",
          "weights"
        ],
        "title": "PortfolioRequest"
      },
      "SeriesPayload": {
        "properties": {
          "series_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Series Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "price",
              "cpi",
              "monthly_rate",
              "wealth",
              "fx"
            ],
            "title": "Kind",
            "default": "price"
          },
          "currency": {
            "type": "string",
            "maxLength": 8,
            "minLength": 3,
            "title": "Currency",
            "default": "TRY"
          },
          "source": {
            "type": "string",
            "maxLength": 1000,
            "title": "Source",
            "default": "user"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic",
            "default": false
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "type": "array",
            "maxItems": 50000,
            "minItems": 1,
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "series_id",
          "data"
        ],
        "title": "SeriesPayload"
      },
      "SpliceRequest": {
        "properties": {
          "left": {
            "$ref": "#/components/schemas/SeriesPayload"
          },
          "right": {
            "$ref": "#/components/schemas/SeriesPayload"
          },
          "splice_date": {
            "type": "string",
            "format": "date",
            "title": "Splice Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "left",
          "right",
          "splice_date"
        ],
        "title": "SpliceRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}